+++ demo/lib/screen/feature/cell_selection_screen.dart
@@ -96,8 +96,8 @@ class _CellSelectionScreenState extends State<CellSelectionScreen> {
             child: Row(
               children: [
                 ElevatedButton(
-                  child: const Text('Show selected cells.'),
                   onPressed: handleSelected,
+                  child: const Text('Show selected cells.'),
                 ),
               ],
             ),