+++ demo/lib/screen/feature/export_screen.dart
@@ -2,7 +2,9 @@ import 'dart:convert';
 
 import 'package:file_saver/file_saver.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:pluto_grid/pluto_grid.dart';
+import 'package:pluto_grid_export/pluto_grid_export.dart' as pluto_grid_export;
 
 import '../../dummy_data/development.dart';
 import '../../widget/pluto_example_button.dart';
@@ -129,10 +131,10 @@ class _ExportScreenState extends State<ExportScreen> {
   @override
   Widget build(BuildContext context) {
     return PlutoExampleScreen(
-      title: 'Export / download as CSV',
-      topTitle: 'Export / download as CSV',
+      title: 'Export / download as PDF or CSV',
+      topTitle: 'Export / download as PDF or CSV',
       topContents: const [
-        Text('You can export grid contents as CSV or TSV'),
+        Text('You can export grid contents as PDF or CSV'),
       ],
       topButtons: [
         PlutoExampleButton(
@@ -171,32 +173,92 @@ class _Header extends StatefulWidget {
 }
 
 class _HeaderState extends State<_Header> {
+  void _printToPdfAndShareOrSave() async {
+    final themeData = pluto_grid_export.ThemeData.withFont(
+      base: pluto_grid_export.Font.ttf(
+        await rootBundle.load('fonts/open_sans/OpenSans-Regular.ttf'),
+      ),
+      bold: pluto_grid_export.Font.ttf(
+        await rootBundle.load('fonts/open_sans/OpenSans-Bold.ttf'),
+      ),
+    );
+
+    var plutoGridPdfExport = pluto_grid_export.PlutoGridDefaultPdfExport(
+      title: "Pluto Grid Sample pdf print",
+      creator: "Pluto Grid Rocks!",
+      format: pluto_grid_export.PdfPageFormat.a4.landscape,
+      themeData: themeData,
+    );
+
+    await pluto_grid_export.Printing.sharePdf(
+        bytes: await plutoGridPdfExport.export(widget.stateManager),
+        filename: plutoGridPdfExport.getFilename());
+  }
+
+  // This doesn't works properly in systems different from Windows.
+  // Disabled for now
+  // void _printToPdfWithDialog() async {
+  //   var originalFormat = PdfPageFormat.a4.landscape;
+  //
+  //   var plutoGridDefaultPdfExport = PlutoGridDefaultPdfExport(
+  //       title: "Pluto Grid Sample pdf print",
+  //       creator: "Pluto Grid Rocks!",
+  //       format: originalFormat);
+  //
+  //   await Printing.layoutPdf(
+  //       format: originalFormat,
+  //       name: plutoGridDefaultPdfExport.getFilename(),
+  //       onLayout: (PdfPageFormat format) async {
+  //         // Update format onLayout
+  //         plutoGridDefaultPdfExport.format = format;
+  //         return plutoGridDefaultPdfExport.export(widget.stateManager);
+  //       });
+  // }
+
   void _defaultExportGridAsCSV() async {
     String title = "pluto_grid_export";
-    var exported =
-        const Utf8Encoder().convert(PlutoExport.exportCSV(widget.stateManager));
+    var exported = const Utf8Encoder().convert(
+        pluto_grid_export.PlutoGridExport.exportCSV(widget.stateManager));
+    await FileSaver.instance.saveFile("$title.csv", exported, ".csv");
+  }
+
+  void _defaultExportGridAsCSVCompatibleWithExcel() async {
+    String title = "pluto_grid_export";
+    var exportCSV =
+        pluto_grid_export.PlutoGridExport.exportCSV(widget.stateManager);
+    var exported = const Utf8Encoder().convert(
+        // FIX Add starting \u{FEFF} / 0xEF, 0xBB, 0xBF
+        // This allows open the file in Excel with proper character interpretation
+        // See https://stackoverflow.com/a/155176
+        '\u{FEFF}$exportCSV');
     await FileSaver.instance.saveFile("$title.csv", exported, ".csv");
   }
 
   void _defaultExportGridAsCSVFakeExcel() async {
     String title = "pluto_grid_export";
-    var exported =
-        const Utf8Encoder().convert(PlutoExport.exportCSV(widget.stateManager));
+    var exportCSV =
+        pluto_grid_export.PlutoGridExport.exportCSV(widget.stateManager);
+    var exported = const Utf8Encoder().convert(
+        // FIX Add starting \u{FEFF} / 0xEF, 0xBB, 0xBF
+        // This allows open the file in Excel with proper character interpretation
+        // See https://stackoverflow.com/a/155176
+        '\u{FEFF}$exportCSV');
     await FileSaver.instance.saveFile("$title.xls", exported, ".xls");
   }
 
-  void _exportGridAsTSV() async {
-    String title = "pluto_grid_export";
-    var exported = const Utf8Encoder().convert(PlutoExport.exportCSV(
-      widget.stateManager,
-      fieldDelimiter: "\t",
-    ));
-    await FileSaver.instance.saveFile("$title.csv", exported, ".csv");
-  }
+  // void _exportGridAsTSV() async {
+  //   String title = "pluto_grid_export";
+  //   var exported = const Utf8Encoder().convert(PlutoGridExport.exportCSV(
+  //     widget.stateManager,
+  //     fieldDelimiter: "\t",
+  //   ));
+  //   await FileSaver.instance.saveFile("$title.csv", exported, ".csv");
+  // }
 
   void _defaultExportGridAsCSVWithSemicolon() async {
     String title = "pluto_grid_export";
-    var exported = const Utf8Encoder().convert(PlutoExport.exportCSV(
+    var exported =
+        const Utf8Encoder().convert(pluto_grid_export.PlutoGridExport.exportCSV(
       widget.stateManager,
       fieldDelimiter: ";",
     ));
@@ -215,17 +277,28 @@ class _HeaderState extends State<_Header> {
             crossAxisAlignment: WrapCrossAlignment.center,
             children: [
               ElevatedButton(
+                  onPressed: _printToPdfAndShareOrSave,
+                  child: const Text("Print to PDF and Share")),
+
+              // TODO This works only under Windows, disabled for now
+              // ElevatedButton(
+              //     onPressed: _printToPdfWithDialog,
+              //     child: const Text("Print PDF with dialog (Windows only)")),
+              ElevatedButton(
                   onPressed: _defaultExportGridAsCSV,
                   child: const Text("Export to CSV")),
               ElevatedButton(
                   onPressed: _defaultExportGridAsCSVWithSemicolon,
                   child: const Text("Export to CSV with Semicolon ';'")),
+              // ElevatedButton(
+              //     onPressed: _exportGridAsTSV,
+              //     child: const Text("Export to TSV (tab separated)")),
               ElevatedButton(
-                  onPressed: _exportGridAsTSV,
-                  child: const Text("Export to TSV (tab separated)")),
+                  onPressed: _defaultExportGridAsCSVCompatibleWithExcel,
+                  child: const Text("UTF-8 CSV compatible with MS Excel")),
               ElevatedButton(
                   onPressed: _defaultExportGridAsCSVFakeExcel,
-                  child: const Text("Export to fake-Excel")),
+                  child: const Text("Fake MS Excel .xls export")),
             ],
           ),
         ),